command and observer testing main
